// FRONTEND/chatbot-admin-panel/src/services/api/schemas/readAllApiClientsParams.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */

export type ReadAllApiClientsParams = {
/**
 * Número de registros a saltar para paginación.
 */
skip?: number;
/**
 * Máximo número de registros a devolver.
 */
limit?: number;
};